+++ pkg/scalers/prometheus_scaler_test.go
@@ -208,3 +208,34 @@ func TestPrometheusScalerExecutePromQuery(t *testing.T) {
 		})
 	}
 }
+
+func TestPrometheusScalerCortexHeader(t *testing.T) {
+	testData := prometheusQromQueryResultTestData{
+		name:           "no values",
+		bodyStr:        `{"data":{"result":[]}}`,
+		responseStatus: http.StatusOK,
+		expectedValue:  0,
+		isError:        false,
+	}
+	cortexOrgValue := "my-org"
+	server := httptest.NewServer(http.HandlerFunc(func(writer http.ResponseWriter, request *http.Request) {
+		reqHeader := request.Header.Get(promCortexHeaderKey)
+		assert.Equal(t, reqHeader, cortexOrgValue)
+		writer.WriteHeader(testData.responseStatus)
+		if _, err := writer.Write([]byte(testData.bodyStr)); err != nil {
+			t.Fatal(err)
+		}
+	}))
+
+	scaler := prometheusScaler{
+		metadata: &prometheusMetadata{
+			serverAddress: server.URL,
+			cortexOrgID:   cortexOrgValue,
+		},
+		httpClient: http.DefaultClient,
+	}
+
+	_, err := scaler.ExecutePromQuery(context.TODO())
+
+	assert.NoError(t, err)
+}